@@ -1,41 +1,42 @@
 {
   "name": "admin",
   "private": true,
-  "version": "2.0.3",
+  "version": "2.2.7",
   "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "build-copy": "tsc && vite build --outDir ../src/templates/admin --emptyOutDir",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@radix-ui/react-switch": "^1.0.3"
+    "@radix-ui/react-switch": "^1.1.3"
   },
   "devDependencies": {
-    "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-toast": "^1.1.5",
-    "@types/react": "^18.2.79",
-    "@types/react-dom": "^18.2.25",
-    "@typescript-eslint/eslint-plugin": "^7.7.0",
-    "@typescript-eslint/parser": "^7.7.0",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "eslint": "^9.0.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.5",
-    "i18next": "^23.11.2",
-    "i18next-browser-languagedetector": "^7.2.1",
-    "lucide-react": "^0.378.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-hook-form": "^7.51.3",
-    "react-i18next": "^14.1.0",
-    "react-router-dom": "^6.22.3",
-    "socket.io-client": "^4.7.5",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.9",
-    "vite-plugin-static-copy": "^1.0.3",
-    "vite-plugin-svgr": "^4.2.0",
-    "zustand": "^4.5.2"
+    "@radix-ui/react-dialog": "^1.1.4",
+    "@radix-ui/react-toast": "^1.2.4",
+    "@types/react": "^19.0.2",
+    "@types/react-dom": "^19.0.2",
+    "@typescript-eslint/eslint-plugin": "^8.18.1",
+    "@typescript-eslint/parser": "^8.18.1",
+    "@vitejs/plugin-react-swc": "^3.7.2",
+    "eslint": "^9.17.0",
+    "eslint-plugin-react-hooks": "^5.1.0",
+    "eslint-plugin-react-refresh": "^0.4.16",
+    "i18next": "^24.2.0",
+    "i18next-browser-languagedetector": "^8.0.2",
+    "lucide-react": "^0.469.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-hook-form": "^7.54.1",
+    "react-i18next": "^15.2.0",
+    "react-router-dom": "^7.0.2",
+    "socket.io-client": "^4.8.1",
+    "typescript": "^5.7.2",
+    "vite": "^6.0.9",
+    "vite-plugin-static-copy": "^2.2.0",
+    "vite-plugin-svgr": "^4.3.0",
+    "zustand": "^5.0.2"
   }
 }
