@@ -30,85 +30,91 @@
     }
   ],
   "dependencies": {
-    "async": "^3.2.5",
-    "axios": "^1.6.8",
-    "clean-css": "^5.3.3",
-    "cookie-parser": "^1.4.6",
-    "cross-spawn": "^7.0.3",
+    "@etherpad/express-session": "^1.18.4",
+    "async": "^3.2.6",
+    "axios": "^1.7.9",
+    "cookie-parser": "^1.4.7",
+    "cross-env": "^7.0.3",
+    "cross-spawn": "^7.0.6",
     "ejs": "^3.1.10",
-    "etherpad-require-kernel": "^1.0.16",
-    "etherpad-yajsml": "0.0.12",
-    "express": "4.19.2",
-    "express-rate-limit": "^7.2.0",
-    "express-session": "npm:@etherpad/express-session@^1.18.2",
+    "esbuild": "^0.25.0",
+    "express": "4.21.2",
+    "express-rate-limit": "^7.5.0",
     "fast-deep-equal": "^3.1.3",
     "find-root": "1.1.0",
-    "formidable": "^3.5.1",
+    "formidable": "^3.5.2",
     "http-errors": "^2.0.0",
-    "jose": "^5.2.4",
+    "jose": "^5.9.6",
     "js-cookie": "^3.0.5",
-    "jsdom": "^24.0.0",
+    "jsdom": "^26.0.0",
     "jsonminify": "0.4.2",
     "jsonwebtoken": "^9.0.2",
     "languages4translatewiki": "0.1.3",
-    "live-plugin-manager": "^0.20.0",
+    "live-plugin-manager": "^1.0.0",
     "lodash.clonedeep": "4.5.0",
     "log4js": "^6.9.1",
+    "lru-cache": "^11.0.2",
     "measured-core": "^2.0.0",
     "mime-types": "^2.1.35",
-    "oidc-provider": "^8.4.5",
-    "openapi-backend": "^5.10.6",
+    "oidc-provider": "^8.6.1",
+    "openapi-backend": "^5.11.1",
     "proxy-addr": "^2.0.7",
-    "rate-limiter-flexible": "^5.0.0",
-    "rehype": "^13.0.1",
-    "rehype-minify-whitespace": "^6.0.0",
-    "resolve": "1.22.8",
+    "rate-limiter-flexible": "^5.0.5",
+    "rehype": "^13.0.2",
+    "rehype-minify-whitespace": "^6.0.2",
+    "resolve": "1.22.10",
+    "rusty-store-kv": "^1.3.1",
     "security": "1.0.0",
-    "semver": "^7.6.0",
-    "socket.io": "^4.7.5",
-    "socket.io-client": "^4.7.5",
-    "superagent": "^9.0.2",
-    "terser": "^5.30.3",
-    "threads": "^1.7.0",
+    "semver": "^7.7.1",
+    "socket.io": "^4.8.1",
+    "socket.io-client": "^4.8.1",
+    "superagent": "10.1.1",
+    "swagger-ui-express": "^5.0.1",
     "tinycon": "0.6.8",
-    "tsx": "^4.7.2",
-    "ueberdb2": "^4.2.63",
-    "underscore": "1.13.6",
+    "tsx": "4.19.2",
+    "ueberdb2": "^5.0.6",
+    "underscore": "1.13.7",
     "unorm": "1.6.0",
-    "wtfnode": "^0.9.2",
-    "lru-cache": "^10.2.0"
+    "wtfnode": "^0.10.0"
   },
   "bin": {
     "etherpad-healthcheck": "../bin/etherpad-healthcheck",
     "etherpad-lite": "node/server.ts"
   },
   "devDependencies": {
-    "@playwright/test": "^1.43.1",
+    "@playwright/test": "^1.49.1",
     "@types/async": "^3.2.24",
     "@types/express": "^4.17.21",
     "@types/formidable": "^3.4.5",
     "@types/http-errors": "^2.0.4",
-    "@types/jsdom": "^21.1.6",
-    "@types/jsonwebtoken": "^9.0.6",
-    "@types/mocha": "^10.0.6",
-    "@types/node": "^20.12.8",
-    "@types/oidc-provider": "^8.4.4",
+    "@types/jquery": "^3.5.32",
+    "@types/js-cookie": "^3.0.6",
+    "@types/jsdom": "^21.1.7",
+    "@types/jsonwebtoken": "^9.0.7",
+    "@types/mime-types": "^2.1.4",
+    "@types/mocha": "^10.0.9",
+    "@types/node": "^22.10.2",
+    "@types/oidc-provider": "^8.5.2",
     "@types/semver": "^7.5.8",
     "@types/sinon": "^17.0.3",
     "@types/supertest": "^6.0.2",
-    "@types/underscore": "^1.11.15",
-    "eslint": "^9.0.0",
+    "@types/swagger-ui-express": "^4.1.7",
+    "@types/underscore": "^1.13.0",
+    "@types/whatwg-mimetype": "^3.0.2",
+    "chokidar": "^4.0.3",
+    "eslint": "^9.17.0",
     "eslint-config-etherpad": "^4.0.4",
     "etherpad-cli-client": "^3.0.2",
-    "mocha": "^10.4.0",
+    "mocha": "^11.0.1",
     "mocha-froth": "^0.2.10",
     "nodeify": "^1.0.1",
     "openapi-schema-validation": "^0.4.2",
-    "set-cookie-parser": "^2.6.0",
-    "sinon": "^17.0.1",
+    "set-cookie-parser": "^2.7.1",
+    "sinon": "^19.0.2",
     "split-grid": "^1.0.11",
     "supertest": "^7.0.0",
-    "typescript": "^5.4.5"
+    "typescript": "^5.7.2",
+    "vitest": "^2.1.8"
   },
   "engines": {
     "node": ">=18.18.2",
