@@ -21,28 +21,30 @@
 
 import {MapArrayType} from "../types/MapType";
 
-const AttributeMap = require('../../static/js/AttributeMap');
+import AttributeMap from '../../static/js/AttributeMap';
 const padManager = require('../db/PadManager');
-const Changeset = require('../../static/js/Changeset');
-const ChatMessage = require('../../static/js/ChatMessage');
-const AttributePool = require('../../static/js/AttributePool');
+import {checkRep, cloneAText, compose, deserializeOps, follow, identity, inverse, makeAText, makeSplice, moveOpsToNewPool, mutateAttributionLines, mutateTextLines, oldLen, prepareForWire, splitAttributionLines, splitTextLines, unpack} from '../../static/js/Changeset';
+import ChatMessage from '../../static/js/ChatMessage';
+import AttributePool from '../../static/js/AttributePool';
 const AttributeManager = require('../../static/js/AttributeManager');
 const authorManager = require('../db/AuthorManager');
-const {padutils} = require('../../static/js/pad_utils');
+import padutils from '../../static/js/pad_utils';
 const readOnlyManager = require('../db/ReadOnlyManager');
 const settings = require('../utils/Settings');
 const securityManager = require('../db/SecurityManager');
-const plugins = require('../../static/js/pluginfw/plugin_defs.js');
+const plugins = require('../../static/js/pluginfw/plugin_defs');
 import log4js from 'log4js';
 const messageLogger = log4js.getLogger('message');
 const accessLogger = log4js.getLogger('access');
-const hooks = require('../../static/js/pluginfw/hooks.js');
+const hooks = require('../../static/js/pluginfw/hooks');
 const stats = require('../stats')
 const assert = require('assert').strict;
 import {RateLimiterMemory} from 'rate-limiter-flexible';
 import {ChangesetRequest, PadUserInfo, SocketClientRequest} from "../types/SocketClientRequest";
 import {APool, AText, PadAuthor, PadType} from "../types/PadType";
 import {ChangeSet} from "../types/ChangeSet";
+import {ChatMessageMessage, ClientReadyMessage, ClientSaveRevisionMessage, ClientSuggestUserName, ClientUserChangesMessage, ClientVarMessage, CustomMessage, PadDeleteMessage, UserNewInfoMessage} from "../../static/js/types/SocketIOMessage";
+import {Builder} from "../../static/js/Builder";
 const webaccess = require('../hooks/express/webaccess');
 const { checkValidRev } = require('../utils/checkValidRev');
 
@@ -209,12 +211,51 @@ exports.handleDisconnect = async (socket:any) => {
   });
 };
 
+
+const handlePadDelete = async (socket: any, padDeleteMessage: PadDeleteMessage) => {
+  const session = sessioninfos[socket.id];
+  if (!session || !session.author || !session.padId) throw new Error('session not ready');
+  if (await padManager.doesPadExist(padDeleteMessage.data.padId)) {
+    const retrievedPad = await padManager.getPad(padDeleteMessage.data.padId)
+    // Only the one doing the first revision can delete the pad, otherwise people could troll a lot
+    const firstContributor = await retrievedPad.getRevisionAuthor(0)
+    if (session.author === firstContributor) {
+      retrievedPad.remove()
+    } else {
+
+      type ShoutMessage = {
+        message: string,
+        sticky: boolean,
+      }
+
+      const messageToShout: ShoutMessage = {
+        message: 'You are not the creator of this pad, so you cannot delete it',
+        sticky: false
+      }
+      const messageToSend = {
+        type: "COLLABROOM",
+        data: {
+          type: "shoutMessage",
+          payload: {
+            message: messageToShout,
+            timestamp: Date.now()
+          }
+        }
+      }
+      socket.emit('shout',
+        messageToSend
+      )
+    }
+  }
+}
+
+
 /**
  * Handles a message from a user
  * @param socket the socket.io Socket object for the client
  * @param message the message from the client
  */
-exports.handleMessage = async (socket:any, message:typeof ChatMessage) => {
+exports.handleMessage = async (socket:any, message: ClientVarMessage) => {
   const env = process.env.NODE_ENV || 'development';
 
   if (env === 'production') {
@@ -348,15 +389,16 @@ exports.handleMessage = async (socket:any, message:typeof ChatMessage) => {
               stats.counter('pendingEdits').inc();
               await padChannels.enqueue(thisSession.padId, {socket, message});
               break;
-            case 'USERINFO_UPDATE': await handleUserInfoUpdate(socket, message); break;
-            case 'CHAT_MESSAGE': await handleChatMessage(socket, message); break;
+            case 'PAD_DELETE': await handlePadDelete(socket, message.data as unknown as PadDeleteMessage); break;
+            case 'USERINFO_UPDATE': await handleUserInfoUpdate(socket, message as unknown as UserNewInfoMessage); break;
+            case 'CHAT_MESSAGE': await handleChatMessage(socket, message as unknown as ChatMessageMessage); break;
             case 'GET_CHAT_MESSAGES': await handleGetChatMessages(socket, message); break;
-            case 'SAVE_REVISION': await handleSaveRevisionMessage(socket, message); break;
+            case 'SAVE_REVISION': await handleSaveRevisionMessage(socket, message as unknown as ClientSaveRevisionMessage); break;
             case 'CLIENT_MESSAGE': {
               const {type} = message.data.payload;
               try {
                 switch (type) {
-                  case 'suggestUserName': handleSuggestUserName(socket, message); break;
+                  case 'suggestUserName': handleSuggestUserName(socket, message as unknown as ClientSuggestUserName); break;
                   default: throw new Error('unknown message type');
                 }
               } catch (err) {
@@ -384,7 +426,7 @@ exports.handleMessage = async (socket:any, message:typeof ChatMessage) => {
  * @param socket the socket.io Socket object for the client
  * @param message the message from the client
  */
-const handleSaveRevisionMessage = async (socket:any, message: string) => {
+const handleSaveRevisionMessage = async (socket:any, message: ClientSaveRevisionMessage) => {
   const {padId, author: authorId} = sessioninfos[socket.id];
   const pad = await padManager.getPad(padId, null, authorId);
   await pad.addSavedRevision(pad.head, authorId);
@@ -397,7 +439,7 @@ const handleSaveRevisionMessage = async (socket:any, message: string) => {
  * @param msg {Object} the message we're sending
  * @param sessionID {string} the socketIO session to which we're sending this message
  */
-exports.handleCustomObjectMessage = (msg: typeof ChatMessage, sessionID: string) => {
+exports.handleCustomObjectMessage = (msg: CustomMessage, sessionID: string) => {
   if (msg.data.type === 'CUSTOM') {
     if (sessionID) {
       // a sessionID is targeted: directly to this sessionID
@@ -432,7 +474,7 @@ exports.handleCustomMessage = (padID: string, msgString:string) => {
  * @param socket the socket.io Socket object for the client
  * @param message the message from the client
  */
-const handleChatMessage = async (socket:any, message: typeof ChatMessage) => {
+const handleChatMessage = async (socket:any, message: ChatMessageMessage) => {
   const chatMessage = ChatMessage.fromObject(message.data.message);
   const {padId, author: authorId} = sessioninfos[socket.id];
   // Don't trust the user-supplied values.
@@ -452,7 +494,7 @@ const handleChatMessage = async (socket:any, message: typeof ChatMessage) => {
  * @param {string} [padId] - The destination pad ID. Deprecated; pass a chat message
  *     object as the first argument and the destination pad ID as the second argument instead.
  */
-exports.sendChatMessageToPadClients = async (mt: typeof ChatMessage|number, puId: string, text:string|null = null, padId:string|null = null) => {
+exports.sendChatMessageToPadClients = async (mt: ChatMessage|number, puId: string, text:string|null = null, padId:string|null = null) => {
   const message = mt instanceof ChatMessage ? mt : new ChatMessage(text, puId, mt);
   padId = mt instanceof ChatMessage ? puId : padId;
   const pad = await padManager.getPad(padId, null, message.authorId);
@@ -499,7 +541,7 @@ const handleGetChatMessages = async (socket:any, {data: {start, end}}:any) => {
  * @param socket the socket.io Socket object for the client
  * @param message the message from the client
  */
-const handleSuggestUserName = (socket:any, message: typeof ChatMessage) => {
+const handleSuggestUserName = (socket:any, message: ClientSuggestUserName) => {
   const {newName, unnamedId} = message.data.payload;
   if (newName == null) throw new Error('missing newName');
   if (unnamedId == null) throw new Error('missing unnamedId');
@@ -519,7 +561,7 @@ const handleSuggestUserName = (socket:any, message: typeof ChatMessage) => {
  * @param socket the socket.io Socket object for the client
  * @param message the message from the client
  */
-const handleUserInfoUpdate = async (socket:any, {data: {userInfo: {name, colorId}}}: PadUserInfo) => {
+const handleUserInfoUpdate = async (socket:any, {data: {userInfo: {name, colorId}}}: UserNewInfoMessage) => {
   if (colorId == null) throw new Error('missing colorId');
   if (!name) name = null;
   const session = sessioninfos[socket.id];
@@ -567,7 +609,9 @@ const handleUserInfoUpdate = async (socket:any, {data: {userInfo: {name, colorId
  * @param socket the socket.io Socket object for the client
  * @param message the message from the client
  */
-const handleUserChanges = async (socket:any, message: typeof ChatMessage) => {
+const handleUserChanges = async (socket:any, message: {
+  data: ClientUserChangesMessage
+}) => {
   // This one's no longer pending, as we're gonna process it now
   stats.counter('pendingEdits').dec();
 
@@ -591,10 +635,10 @@ const handleUserChanges = async (socket:any, message: typeof ChatMessage) => {
     const pad = await padManager.getPad(thisSession.padId, null, thisSession.author);
 
     // Verify that the changeset has valid syntax and is in canonical form
-    Changeset.checkRep(changeset);
+    checkRep(changeset);
 
     // Validate all added 'author' attribs to be the same value as the current user
-    for (const op of Changeset.deserializeOps(Changeset.unpack(changeset).ops)) {
+    for (const op of deserializeOps(unpack(changeset).ops)) {
       // + can add text with attribs
       // = can change or add attribs
       // - can have attribs, but they are discarded and don't show up in the attribs -
@@ -613,7 +657,7 @@ const handleUserChanges = async (socket:any, message: typeof ChatMessage) => {
     // ex. adoptChangesetAttribs
 
     // Afaik, it copies the new attributes from the changeset, to the global Attribute Pool
-    let rebasedChangeset = Changeset.moveOpsToNewPool(changeset, wireApool, pad.pool);
+    let rebasedChangeset = moveOpsToNewPool(changeset, wireApool, pad.pool);
 
     // ex. applyUserChanges
     let r = baseRev;
@@ -626,21 +670,21 @@ const handleUserChanges = async (socket:any, message: typeof ChatMessage) => {
       const {changeset: c, meta: {author: authorId}} = await pad.getRevision(r);
       if (changeset === c && thisSession.author === authorId) {
         // Assume this is a retransmission of an already applied changeset.
-        rebasedChangeset = Changeset.identity(Changeset.unpack(changeset).oldLen);
+        rebasedChangeset = identity(unpack(changeset).oldLen);
       }
       // At this point, both "c" (from the pad) and "changeset" (from the
       // client) are relative to revision r - 1. The follow function
       // rebases "changeset" so that it is relative to revision r
       // and can be applied after "c".
-      rebasedChangeset = Changeset.follow(c, rebasedChangeset, false, pad.pool);
+      rebasedChangeset = follow(c, rebasedChangeset, false, pad.pool);
     }
 
     const prevText = pad.text();
 
-    if (Changeset.oldLen(rebasedChangeset) !== prevText.length) {
+    if (oldLen(rebasedChangeset) !== prevText.length) {
       throw new Error(
           `Can't apply changeset ${rebasedChangeset} with oldLen ` +
-          `${Changeset.oldLen(rebasedChangeset)} to document of length ${prevText.length}`);
+          `${oldLen(rebasedChangeset)} to document of length ${prevText.length}`);
     }
 
     const newRev = await pad.appendRevision(rebasedChangeset, thisSession.author);
@@ -655,7 +699,7 @@ const handleUserChanges = async (socket:any, message: typeof ChatMessage) => {
 
     // Make sure the pad always ends with an empty line.
     if (pad.text().lastIndexOf('\n') !== pad.text().length - 1) {
-      const nlChangeset = Changeset.makeSplice(pad.text(), pad.text().length - 1, 0, '\n');
+      const nlChangeset = makeSplice(pad.text(), pad.text().length - 1, 0, '\n');
       await pad.appendRevision(nlChangeset, thisSession.author);
     }
 
@@ -710,7 +754,7 @@ exports.updatePadClients = async (pad: PadType) => {
       const revChangeset = revision.changeset;
       const currentTime = revision.meta.timestamp;
 
-      const forWire = Changeset.prepareForWire(revChangeset, pad.pool);
+      const forWire = prepareForWire(revChangeset, pad.pool);
       const msg = {
         type: 'COLLABROOM',
         data: {
@@ -738,14 +782,14 @@ exports.updatePadClients = async (pad: PadType) => {
 /**
  * Copied from the Etherpad Source Code. Don't know what this method does excatly...
  */
-const _correctMarkersInPad = (atext: AText, apool: APool) => {
+const _correctMarkersInPad = (atext: AText, apool: AttributePool) => {
   const text = atext.text;
 
   // collect char positions of line markers (e.g. bullets) in new atext
   // that aren't at the start of a line
   const badMarkers = [];
   let offset = 0;
-  for (const op of Changeset.deserializeOps(atext.attribs)) {
+  for (const op of deserializeOps(atext.attribs)) {
     const attribs = AttributeMap.fromString(op.attribs, apool);
     const hasMarker = AttributeManager.lineAttributes.some((a: string) => attribs.has(a));
     if (hasMarker) {
@@ -767,7 +811,7 @@ const _correctMarkersInPad = (atext: AText, apool: APool) => {
   // create changeset that removes these bad markers
   offset = 0;
 
-  const builder = Changeset.builder(text.length);
+  const builder = new Builder(text.length);
 
   badMarkers.forEach((pos) => {
     builder.keepText(text.substring(offset, pos));
@@ -785,7 +829,7 @@ const _correctMarkersInPad = (atext: AText, apool: APool) => {
  * @param socket the socket.io Socket object for the client
  * @param message the message from the client
  */
-const handleClientReady = async (socket:any, message: typeof ChatMessage) => {
+const handleClientReady = async (socket:any, message: ClientReadyMessage) => {
   const sessionInfo = sessioninfos[socket.id];
   if (sessionInfo == null) throw new Error('client disconnected');
   assert(sessionInfo.author);
@@ -793,8 +837,9 @@ const handleClientReady = async (socket:any, message: typeof ChatMessage) => {
   await hooks.aCallAll('clientReady', message); // Deprecated due to awkward context.
 
   let {colorId: authorColorId, name: authorName} = message.userInfo || {};
-  if (authorColorId && !/^#(?:[0-9A-F]{3}){1,2}$/i.test(authorColorId)) {
+  if (authorColorId && !/^#(?:[0-9A-F]{3}){1,2}$/i.test(authorColorId as string)) {
     messageLogger.warn(`Ignoring invalid colorId in CLIENT_READY message: ${authorColorId}`);
+    // @ts-ignore
     authorColorId = null;
   }
   await Promise.all([
@@ -872,7 +917,7 @@ const handleClientReady = async (socket:any, message: typeof ChatMessage) => {
     const revisionsNeeded = [];
     const changesets:MapArrayType<any> = {};
 
-    let startNum = message.client_rev + 1;
+    let startNum = message.client_rev! + 1;
     let endNum = pad.getHeadRevisionNumber() + 1;
 
     const headNum = pad.getHeadRevisionNumber();
@@ -901,7 +946,7 @@ const handleClientReady = async (socket:any, message: typeof ChatMessage) => {
 
     // return pending changesets
     for (const r of revisionsNeeded) {
-      const forWire = Changeset.prepareForWire(changesets[r].changeset, pad.pool);
+      const forWire = prepareForWire(changesets[r].changeset, pad.pool);
       const wireMsg = {type: 'COLLABROOM',
         data: {type: 'CLIENT_RECONNECT',
           headRev: pad.getHeadRevisionNumber(),
@@ -926,8 +971,8 @@ const handleClientReady = async (socket:any, message: typeof ChatMessage) => {
     let apool;
     // prepare all values for the wire, there's a chance that this throws, if the pad is corrupted
     try {
-      atext = Changeset.cloneAText(pad.atext);
-      const attribsForWire = Changeset.prepareForWire(atext.attribs, pad.pool);
+      atext = cloneAText(pad.atext);
+      const attribsForWire = prepareForWire(atext.attribs, pad.pool);
       apool = attribsForWire.pool.toJsonable();
       atext.attribs = attribsForWire.translated;
     } catch (e:any) {
@@ -996,7 +1041,8 @@ const handleClientReady = async (socket:any, message: typeof ChatMessage) => {
         percentageToScrollWhenUserPressesArrowUp:
             settings.scrollWhenFocusLineIsOutOfViewport.percentageToScrollWhenUserPressesArrowUp,
       },
-      initialChangesets: [], // FIXME: REMOVE THIS SHIT
+      initialChangesets: [], // FIXME: REMOVE THIS SHIT,
+      mode: process.env.NODE_ENV
     };
 
     // Add a username to the clientVars if one avaiable
@@ -1141,13 +1187,13 @@ const getChangesetInfo = async (pad: PadType, startNum: number, endNum:number, g
     getPadLines(pad, startNum - 1),
     // Get all needed composite Changesets.
     ...compositesChangesetNeeded.map(async (item) => {
-      const changeset = await composePadChangesets(pad, item.start, item.end);
+      const changeset = await exports.composePadChangesets(pad, item.start, item.end);
       composedChangesets[`${item.start}/${item.end}`] = changeset;
     }),
     // Get all needed revision Dates.
     ...revTimesNeeded.map(async (revNum) => {
       const revDate = await pad.getRevisionDate(revNum);
-      revisionDate[revNum] = Math.floor(revDate / 1000);
+      revisionDate[revNum] = revDate;
     }),
   ]);
 
@@ -1162,13 +1208,13 @@ const getChangesetInfo = async (pad: PadType, startNum: number, endNum:number, g
     if (compositeEnd > endNum || compositeEnd > headRevision + 1) break;
 
     const forwards = composedChangesets[`${compositeStart}/${compositeEnd}`];
-    const backwards = Changeset.inverse(forwards, lines.textlines, lines.alines, pad.apool());
+    const backwards = inverse(forwards, lines.textlines, lines.alines, pad.apool());
 
-    Changeset.mutateAttributionLines(forwards, lines.alines, pad.apool());
-    Changeset.mutateTextLines(forwards, lines.textlines);
+    mutateAttributionLines(forwards, lines.alines, pad.apool());
+    mutateTextLines(forwards, lines.textlines);
 
-    const forwards2 = Changeset.moveOpsToNewPool(forwards, pad.apool(), apool);
-    const backwards2 = Changeset.moveOpsToNewPool(backwards, pad.apool(), apool);
+    const forwards2 = moveOpsToNewPool(forwards, pad.apool(), apool);
+    const backwards2 = moveOpsToNewPool(backwards, pad.apool(), apool);
 
     const t1 = (compositeStart === 0) ? revisionDate[0] : revisionDate[compositeStart - 1];
     const t2 = revisionDate[compositeEnd - 1];
@@ -1194,12 +1240,12 @@ const getPadLines = async (pad: PadType, revNum: number) => {
   if (revNum >= 0) {
     atext = await pad.getInternalRevisionAText(revNum);
   } else {
-    atext = Changeset.makeAText('\n');
+    atext = makeAText('\n');
   }
 
   return {
-    textlines: Changeset.splitTextLines(atext.text),
-    alines: Changeset.splitAttributionLines(atext.attribs, atext.text),
+    textlines: splitTextLines(atext.text),
+    alines: splitAttributionLines(atext.attribs, atext.text),
   };
 };
 
@@ -1207,7 +1253,7 @@ const getPadLines = async (pad: PadType, revNum: number) => {
  * Tries to rebuild the composePadChangeset function of the original Etherpad
  * https://github.com/ether/pad/blob/master/etherpad/src/etherpad/control/pad/pad_changeset_control.js#L241
  */
-const composePadChangesets = async (pad: PadType, startNum: number, endNum: number) => {
+exports.composePadChangesets = async (pad: PadType, startNum: number, endNum: number) => {
   // fetch all changesets we need
   const headNum = pad.getHeadRevisionNumber();
   endNum = Math.min(endNum, headNum + 1);
@@ -1234,7 +1280,7 @@ const composePadChangesets = async (pad: PadType, startNum: number, endNum: numb
 
     for (r = startNum + 1; r < endNum; r++) {
       const cs = changesets[r];
-      changeset = Changeset.compose(changeset, cs, pool);
+      changeset = compose(changeset as string, cs as string, pool);
     }
     return changeset;
   } catch (e) {
