@@ -7,8 +7,8 @@ const languages = require('languages4translatewiki');
 const fs = require('fs');
 const path = require('path');
 const _ = require('underscore');
-const pluginDefs = require('../../static/js/pluginfw/plugin_defs.js');
-const existsSync = require('../utils/path_exists');
+const pluginDefs = require('../../static/js/pluginfw/plugin_defs');
+import existsSync from '../utils/path_exists';
 const settings = require('../utils/Settings');
 
 // returns all existing messages merged together and grouped by langcode
