@@ -19,8 +19,9 @@
  * limitations under the License.
  */
 
-const AttributeMap = require('../../static/js/AttributeMap');
-const Changeset = require('../../static/js/Changeset');
+import AttributeMap from '../../static/js/AttributeMap';
+import AttributePool from "../../static/js/AttributePool";
+import {deserializeOps, splitAttributionLines, subattribution} from '../../static/js/Changeset';
 const { checkValidRev } = require('./checkValidRev');
 
 /*
@@ -30,7 +31,7 @@ exports.getPadPlainText = (pad: { getInternalRevisionAText: (arg0: any) => any; 
   const _analyzeLine = exports._analyzeLine;
   const atext = ((revNum !== undefined) ? pad.getInternalRevisionAText(checkValidRev(revNum)) : pad.atext);
   const textLines = atext.text.slice(0, -1).split('\n');
-  const attribLines = Changeset.splitAttributionLines(atext.attribs, atext.text);
+  const attribLines = splitAttributionLines(atext.attribs, atext.text);
   const apool = pad.pool;
 
   const pieces = [];
@@ -51,14 +52,14 @@ type LineModel = {
   [id:string]:string|number|LineModel
 }
 
-exports._analyzeLine = (text:string, aline: LineModel, apool: Function) => {
+exports._analyzeLine = (text:string, aline: string, apool: AttributePool) => {
   const line: LineModel = {};
 
   // identify list
   let lineMarker = 0;
   line.listLevel = 0;
   if (aline) {
-    const [op] = Changeset.deserializeOps(aline);
+    const [op] = deserializeOps(aline);
     if (op != null) {
       const attribs = AttributeMap.fromString(op.attribs, apool);
       let listType = attribs.get('list');
@@ -78,7 +79,7 @@ exports._analyzeLine = (text:string, aline: LineModel, apool: Function) => {
   }
   if (lineMarker) {
     line.text = text.substring(1);
-    line.aline = Changeset.subattribution(aline, 1);
+    line.aline = subattribution(aline, 1);
   } else {
     line.text = text;
     line.aline = aline;
