@@ -6,7 +6,7 @@ import {useDebounce} from "../utils/useDebounce.ts";
 import {determineSorting} from "../utils/sorting.ts";
 import * as Dialog from "@radix-ui/react-dialog";
 import {IconButton} from "../components/IconButton.tsx";
-import {ChevronLeft, ChevronRight, Eye, Trash2} from "lucide-react";
+import {ChevronLeft, ChevronRight, Eye, Trash2, FileStack} from "lucide-react";
 import {SearchField} from "../components/SearchField.tsx";
 
 export const PadPage = ()=>{
@@ -23,10 +23,11 @@ export const PadPage = ()=>{
     const pads = useStore(state=>state.pads)
     const [currentPage, setCurrentPage] = useState<number>(0)
     const [deleteDialog, setDeleteDialog] = useState<boolean>(false)
+    const [errorText, setErrorText] = useState<string|null>(null)
     const [padToDelete, setPadToDelete] = useState<string>('')
     const pages = useMemo(()=>{
         if(!pads){
-            return [0]
+            return 0;
         }
 
         return Math.ceil(pads!.total / searchParams.limit)
@@ -68,12 +69,35 @@ export const PadPage = ()=>{
                 results: newPads
             })
         })
+
+        settingsSocket.on('results:cleanupPadRevisions', (data)=>{
+          let newPads = useStore.getState().pads?.results ?? []
+
+          if (data.error) {
+            setErrorText(data.error)
+            return
+          }
+
+          newPads.forEach((pad)=>{
+            if (pad.padName === data.padId) {
+              pad.revisionNumber = data.keepRevisions
+            }
+          })
+
+          useStore.getState().setPads({
+            results: newPads,
+            total: useStore.getState().pads!.total
+          })
+        })
     }, [settingsSocket, pads]);
 
     const deletePad = (padID: string)=>{
         settingsSocket?.emit('deletePad', padID)
     }
 
+    const cleanupPad = (padID: string)=>{
+        settingsSocket?.emit('cleanupPadRevisions', padID)
+    }
 
 
     return <div>
@@ -100,11 +124,26 @@ export const PadPage = ()=>{
             </Dialog.Content>
         </Dialog.Portal>
         </Dialog.Root>
+        <Dialog.Root open={errorText !== null}>
+          <Dialog.Portal>
+            <Dialog.Overlay className="dialog-confirm-overlay"/>
+            <Dialog.Content className="dialog-confirm-content">
+              <div>
+                <div>Error occured: {errorText}</div>
+                <div className="settings-button-bar">
+                  <button onClick={() => {
+                    setErrorText(null)
+                  }}>OK</button>
+                </div>
+              </div>
+            </Dialog.Content>
+          </Dialog.Portal>
+        </Dialog.Root>
         <h1><Trans i18nKey="ep_admin_pads:ep_adminpads2_manage-pads"/></h1>
         <SearchField value={searchTerm} onChange={v=>setSearchTerm(v.target.value)} placeholder={t('ep_admin_pads:ep_adminpads2_search-heading')}/>
         <table>
             <thead>
-            <tr>
+            <tr className="search-pads">
                 <th className={determineSorting(searchParams.sortBy, searchParams.ascending, 'padName')} onClick={()=>{
                     setSearchParams({
                         ...searchParams,
@@ -112,17 +151,17 @@ export const PadPage = ()=>{
                         ascending: !searchParams.ascending
                     })
                 }}><Trans i18nKey="ep_admin_pads:ep_adminpads2_padname"/></th>
-                <th className={determineSorting(searchParams.sortBy, searchParams.ascending, 'lastEdited')} onClick={()=>{
-                    setSearchParams({
-                        ...searchParams,
-                        sortBy: 'lastEdited',
-                        ascending: !searchParams.ascending
-                    })
-                }}><Trans i18nKey="ep_admin_pads:ep_adminpads2_pad-user-count"/></th>
                 <th className={determineSorting(searchParams.sortBy, searchParams.ascending, 'userCount')} onClick={()=>{
                     setSearchParams({
                         ...searchParams,
                         sortBy: 'userCount',
+                        ascending: !searchParams.ascending
+                    })
+                }}><Trans i18nKey="ep_admin_pads:ep_adminpads2_pad-user-count"/></th>
+                <th className={determineSorting(searchParams.sortBy, searchParams.ascending, 'lastEdited')} onClick={()=>{
+                    setSearchParams({
+                        ...searchParams,
+                        sortBy: 'lastEdited',
                         ascending: !searchParams.ascending
                     })
                 }}><Trans i18nKey="ep_admin_pads:ep_adminpads2_last-edited"/></th>
@@ -136,7 +175,7 @@ export const PadPage = ()=>{
                 <th><Trans i18nKey="ep_admin_pads:ep_adminpads2_action"/></th>
             </tr>
             </thead>
-            <tbody>
+            <tbody className="search-pads-body">
             {
                 pads?.results?.map((pad)=>{
                     return <tr key={pad.padName}>
@@ -149,6 +188,9 @@ export const PadPage = ()=>{
                                 <IconButton icon={<Trash2/>} title={<Trans i18nKey="ep_admin_pads:ep_adminpads2_delete.value"/>} onClick={()=>{
                                     setPadToDelete(pad.padName)
                                     setDeleteDialog(true)
+                                }}/>
+                                <IconButton icon={<FileStack/>} title={<Trans i18nKey="ep_admin_pads:ep_adminpads2_cleanup"/>} onClick={()=>{
+                                  cleanupPad(pad.padName)
                                 }}/>
                                 <IconButton icon={<Eye/>} title="view" onClick={()=>window.open(`/p/${pad.padName}`, '_blank')}/>
                             </div>
@@ -166,11 +208,12 @@ export const PadPage = ()=>{
                         offset: (Number(currentPage)-1)*searchParams.limit})
             }}><ChevronLeft/><span>Previous Page</span></button>
             <span>{currentPage+1} out of {pages}</span>
-            <button disabled={pages == currentPage+1} onClick={()=>{
-                setCurrentPage(currentPage+1)
+            <button disabled={pages == 0 || pages == currentPage+1} onClick={()=>{
+              const newCurrentPage = currentPage+1
+                setCurrentPage(newCurrentPage)
                 setSearchParams({
                     ...searchParams,
-                    offset: (Number(currentPage)-1)*searchParams.limit
+                    offset: (Number(newCurrentPage))*searchParams.limit
                 })
             }}><span>Next Page</span><ChevronRight/></button>
         </div>
