@@ -22,7 +22,7 @@
 
 const ejs = require('ejs');
 const fs = require('fs');
-const hooks = require('../../static/js/pluginfw/hooks.js');
+const hooks = require('../../static/js/pluginfw/hooks');
 const path = require('path');
 const resolve = require('resolve');
 const settings = require('../utils/Settings');
