@@ -3,6 +3,7 @@
 
 import {PadQueryResult, PadSearchQuery} from "../../types/PadSearchQuery";
 import {PadType} from "../../types/PadType";
+import log4js from 'log4js';
 
 const eejs = require('../../eejs');
 const fsp = require('fs').promises;
@@ -12,9 +13,11 @@ const settings = require('../../utils/Settings');
 const UpdateCheck = require('../../utils/UpdateCheck');
 const padManager = require('../../db/PadManager');
 const api = require('../../db/API');
+const cleanup = require('../../utils/Cleanup');
 
 
 const queryPadLimit = 12;
+const logger = log4js.getLogger('adminSettings');
 
 
 exports.socketio = (hookName: string, {io}: any) => {
@@ -28,7 +31,7 @@ exports.socketio = (hookName: string, {io}: any) => {
             try {
                 data = await fsp.readFile(settings.settingsFilename, 'utf8');
             } catch (err) {
-                return console.log(err);
+                return logger.error(`Error loading settings: ${err}`);
             }
             // if showSettingsInAdminPage is set to false, then return NOT_ALLOWED in the result
             if (settings.showSettingsInAdminPage === false) {
@@ -39,8 +42,12 @@ exports.socketio = (hookName: string, {io}: any) => {
         });
 
         socket.on('saveSettings', async (newSettings: string) => {
-            console.log('Admin request to save settings through a socket on /admin/settings');
-            await fsp.writeFile(settings.settingsFilename, newSettings);
+            logger.info('Admin request to save settings through a socket on /admin/settings');
+            try {
+                await fsp.writeFile(settings.settingsFilename, newSettings);
+            } catch (err) {
+                logger.error(`Error saving settings: ${err}`);
+            }
             socket.emit('saveprogress', 'saved');
         });
 
@@ -122,6 +129,7 @@ exports.socketio = (hookName: string, {io}: any) => {
                 maxResult = 0;
             }
 
+            // Reset to default values if out of bounds
             if (query.offset && query.offset < 0) {
                 query.offset = 0;
             } else if (query.offset > maxResult) {
@@ -129,10 +137,13 @@ exports.socketio = (hookName: string, {io}: any) => {
             }
 
             if (query.limit && query.limit < 0) {
+              // Too small
                 query.limit = 0;
             } else if (query.limit > queryPadLimit) {
+              // Too big
                 query.limit = queryPadLimit;
             }
+
 
             if (query.sortBy === 'padName') {
                 result = result.sort((a, b) => {
@@ -154,53 +165,78 @@ exports.socketio = (hookName: string, {io}: any) => {
                         revisionNumber
                     }
                 }));
-            } else {
+            } else if (query.sortBy === "revisionNumber") {
                 const currentWinners: PadQueryResult[] = []
-                let queryOffsetCounter = 0
+                const padMapping = [] as {padId: string, revisionNumber: number}[]
                 for (let res of result) {
-
                     const pad = await padManager.getPad(res);
-                    const padType = {
-                        padName: res,
-                        lastEdited: await pad.getLastEdit(),
-                        userCount: api.padUsersCount(res).padUsersCount,
-                        revisionNumber: pad.getHeadRevisionNumber()
-                    };
-
-                    if (currentWinners.length < query.limit) {
-                        if (queryOffsetCounter < query.offset) {
-                            queryOffsetCounter++
-                            continue
-                        }
-                        currentWinners.push({
-                            padName: res,
-                            lastEdited: await pad.getLastEdit(),
-                            userCount: api.padUsersCount(res).padUsersCount,
-                            revisionNumber: pad.getHeadRevisionNumber()
-                        })
-                    } else {
-                        // Kick out worst pad and replace by current pad
-                        let worstPad = currentWinners.sort((a, b) => {
-                            if (a[query.sortBy] < b[query.sortBy]) return query.ascending ? -1 : 1;
-                            if (a[query.sortBy] > b[query.sortBy]) return query.ascending ? 1 : -1;
-                            return 0;
-                        })
-                        if (worstPad[0] && worstPad[0][query.sortBy] < padType[query.sortBy]) {
-                            if (queryOffsetCounter < query.offset) {
-                                queryOffsetCounter++
-                                continue
-                            }
-                            currentWinners.splice(currentWinners.indexOf(worstPad[0]), 1)
-                            currentWinners.push({
-                                padName: res,
-                                lastEdited: await pad.getLastEdit(),
-                                userCount: api.padUsersCount(res).padUsersCount,
-                                revisionNumber: pad.getHeadRevisionNumber()
-                            })
-                        }
-                    }
+                    const revisionNumber = pad.getHeadRevisionNumber()
+                    padMapping.push({padId: res, revisionNumber})
                 }
-                data.results = currentWinners;
+                padMapping.sort((a, b) => {
+                    if (a.revisionNumber < b.revisionNumber) return query.ascending ? -1 : 1;
+                    if (a.revisionNumber > b.revisionNumber) return query.ascending ? 1 : -1;
+                    return 0;
+                })
+
+              for (const padRetrieval of padMapping.slice(query.offset, query.offset + query.limit)) {
+                let pad = await padManager.getPad(padRetrieval.padId);
+                currentWinners.push({
+                  padName: padRetrieval.padId,
+                  lastEdited: await pad.getLastEdit(),
+                  userCount: api.padUsersCount(pad.padName).padUsersCount,
+                  revisionNumber: padRetrieval.revisionNumber
+                })
+              }
+
+              data.results = currentWinners;
+            } else if (query.sortBy === "userCount") {
+              const currentWinners: PadQueryResult[] = []
+              const padMapping = [] as {padId: string, userCount: number}[]
+              for (let res of result) {
+                const userCount = api.padUsersCount(res).padUsersCount
+                padMapping.push({padId: res, userCount})
+              }
+              padMapping.sort((a, b) => {
+                if (a.userCount < b.userCount) return query.ascending ? -1 : 1;
+                if (a.userCount > b.userCount) return query.ascending ? 1 : -1;
+                return 0;
+              })
+
+              for (const padRetrieval of padMapping.slice(query.offset, query.offset + query.limit)) {
+                let pad = await padManager.getPad(padRetrieval.padId);
+                currentWinners.push({
+                  padName: padRetrieval.padId,
+                  lastEdited: await pad.getLastEdit(),
+                  userCount: padRetrieval.userCount,
+                  revisionNumber: pad.getHeadRevisionNumber()
+                })
+              }
+              data.results = currentWinners;
+            } else if (query.sortBy === "lastEdited") {
+              const currentWinners: PadQueryResult[] = []
+              const padMapping = [] as {padId: string, lastEdited: string}[]
+              for (let res of result) {
+                const pad = await padManager.getPad(res);
+                const lastEdited = await pad.getLastEdit();
+                padMapping.push({padId: res, lastEdited})
+              }
+              padMapping.sort((a, b) => {
+                if (a.lastEdited < b.lastEdited) return query.ascending ? -1 : 1;
+                if (a.lastEdited > b.lastEdited) return query.ascending ? 1 : -1;
+                return 0;
+              })
+
+              for (const padRetrieval of padMapping.slice(query.offset, query.offset + query.limit)) {
+                let pad = await padManager.getPad(padRetrieval.padId);
+                currentWinners.push({
+                  padName: padRetrieval.padId,
+                  lastEdited: padRetrieval.lastEdited,
+                  userCount: api.padUsersCount(pad.padName).padUsersCount,
+                  revisionNumber: pad.getHeadRevisionNumber()
+                })
+              }
+              data.results = currentWinners;
             }
 
             socket.emit('results:padLoad', data);
@@ -210,14 +246,49 @@ exports.socketio = (hookName: string, {io}: any) => {
         socket.on('deletePad', async (padId: string) => {
             const padExists = await padManager.doesPadExists(padId);
             if (padExists) {
+                logger.info(`Deleting pad: ${padId}`);
                 const pad = await padManager.getPad(padId);
                 await pad.remove();
                 socket.emit('results:deletePad', padId);
             }
         })
 
+        socket.on('cleanupPadRevisions', async (padId: string) => {
+          if (!settings.cleanup.enabled) {
+            socket.emit('results:cleanupPadRevisions', {
+              error: 'Cleanup disabled. Enable cleanup in settings.json: cleanup.enabled => true',
+            });
+            return;
+          }
+
+          const padExists = await padManager.doesPadExists(padId);
+          if (padExists) {
+            logger.info(`Cleanup pad revisions: ${padId}`);
+            try {
+              const result = await cleanup.deleteRevisions(padId, settings.cleanup.keepRevisions)
+              if (result) {
+                socket.emit('results:cleanupPadRevisions', {
+                  padId: padId,
+                  keepRevisions: settings.cleanup.keepRevisions,
+                });
+                logger.info('successful cleaned up pad: ', padId)
+              } else {
+                socket.emit('results:cleanupPadRevisions', {
+                  error: 'Error cleaning up pad',
+                });
+              }
+            } catch (err: any) {
+              logger.error(`Error in pad ${padId}: ${err.stack || err}`);
+              socket.emit('results:cleanupPadRevisions', {
+                error: err.toString(),
+              });
+              return;
+            }
+          }
+        })
+
         socket.on('restartServer', async () => {
-            console.log('Admin request to restart server through a socket on /admin/settings');
+            logger.info('Admin request to restart server through a socket on /admin/settings');
             settings.reloadSettings();
             await plugins.update();
             await hooks.aCallAll('loadSettings', {settings});
@@ -230,4 +301,3 @@ exports.socketio = (hookName: string, {io}: any) => {
 const searchPad = async (query: PadSearchQuery) => {
 
 }
-
