@@ -1,10 +1,11 @@
 import {MapArrayType} from "./MapType";
+import AttributePool from "../../static/js/AttributePool";
 
 export type PadType = {
     id: string,
-    apool: ()=>APool,
+    apool: ()=>AttributePool,
     atext: AText,
-    pool: APool,
+    pool: AttributePool,
     getInternalRevisionAText: (text:number|string)=>Promise<AText>,
     getValidRevisionRange: (fromRev: string, toRev: string)=>PadRange,
     getRevisionAuthor: (rev: number)=>Promise<string>,
@@ -35,6 +36,7 @@ export type APool = {
     clone: ()=>APool,
     check: ()=>Promise<void>,
     eachAttrib: (callback: (key: string, value: any)=>void)=>void,
+    getAttrib: (key: number)=>any,
 }
 
 
