@@ -17,7 +17,7 @@ export default defineConfig({
     expect: { timeout: defaultExpectTimeout },
     timeout: defaultTestTimeout,
     retries: 2,
-    workers: 20,
+    workers: 5,
     /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
     use: {
         /* Base URL to use in actions like `await page.goto('/')`. */
